# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gotenberg
  version: 7.x
  license:
    name: MIT
    url: 'https://github.com/gotenberg/gotenberg/blob/main/LICENSE'
  contact:
    url: 'https://github.com/gotenberg/gotenberg'
  description: >-
    A Docker-powered stateless API for PDF files.
...
servers:
  - url: 'http://localhost:3000'
    description: Local server with the default Docker image and port
...
paths:
  /forms/chromium/convert/url:
    post:
      tags:
        - chromium
      summary: Convert the contents of a given URL to PDF
      externalDocs:
        url: https://gotenberg.dev/docs/modules/chromium
      description: >-
        Send a URL in your API request via the `url` form field
        Send a remote URL in your API request via the `remoteURL` parameter, and
        get the resulting PDF file. The API will fetch the given URL and render
        the page to PDF using the underlying headless Chrome instance.
        You can optionally include `header.html` and `footer.html` files as part of the request as well.
        See externalDocs for more details.
      parameters:
        - $ref: '#/components/parameters/OutputFilename'
        - $ref: '#/components/parameters/Trace'
      requestBody:
        required: true
        description: >-
          The request must be  `multipart/form-data` that includes a `url` form field.
          The API uses a headless Chrome instance to do the conversion, therefore print
          parameter such as margins and paper size are also accepted as optional parameters.
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/URLConvertRequestBody'
            examples: { }
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulPDF'
        '400':
          description: Bad Request
  /forms/chromium/convert/html:
    post:
      tags:
        - chromium
      summary: Convert a given HTML file to PDF
      externalDocs:
        url: https://gotenberg.dev/docs/modules/chromium
      description: >-
        Send an HTML file called `index.html` as a multipart form request, and
        get the resulting PDF file. You can optionally include `header.html` and
        `footer.html` files as part of the request as well.
        See externalDocs for more details.
      parameters:
        - $ref: '#/components/parameters/OutputFilename'
        - $ref: '#/components/parameters/Trace'
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/HTMLConvertRequestBody'
        description: >-
          The request body must have an `index.html` file in the `files` array,
          as well as all the referred resources on the same level as the
          `index.html` file. The request can also include `header.html` and
          `footer.html`, given the limitations in the API description above.
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulPDF'
        '400':
          description: Bad Request
  /forms/chromium/convert/markdown:
    post:
      tags:
        - chromium
      summary: Convert a Markdown file to PDF
      externalDocs:
        url: https://gotenberg.dev/docs/modules/chromium
      description: >-
        Accepts an HTML file called `index.html` plus markdown files as a multipart
        form request and embeds the markdown files into the HTML file using the Golang template
        function `toHTML`.
        The API will convert the markdown to HTML and embed it into your `index.html` file,
        then render the resulting page. You can include your own styling and more in your HTML file.
        Refer to the HTML conversion page for all the options you can use when converting
        Markdown documents as well. You can optionally include `header.html` and
        `footer.html` files as part of the request as well.
        See externalDocs for more details.
      parameters:
        - $ref: '#/components/parameters/OutputFilename'
        - $ref: '#/components/parameters/Trace'
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MarkdownConvertRequestBody'
        description: >-
          The request body must have an `index.html` file in the `files` array,
          as well as all the referred markdown resources on the same level as
          the `index.html` file.
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulPDF'
        '400':
          description: Bad Request
...
components:
  schemas:
    HTMLConvertRequestBody:
      title: HTML Conversion Request Body
      allOf:
      - type: object
        properties:
          files:
            type: array
            description: >-
              List of HTML files to be converted to PDF. An `index.html` file is
              required, and any other resources that are referenced through the
              HTML file must be included as well. All the referenced files must be
              on the same level as the `index.html` file.
            items:
              type: string
              format: binary
          marginTop:
            type: number
            example: 0
            default: 1
            description: Top margin for the page in inches.
          marginBottom:
            type: number
            example: 0
            default: 1
            description: Bottom margin for the page in inches.
          marginLeft:
            type: number
            example: 0
            default: 1
            description: Left margin for the page in inches.
          marginRight:
            type: number
            example: 0
            default: 1
            description: Right margin for the page in inches.
          paperWidth:
            type: number
            example: 8.27
            description: >-
              Paper width to be used while rendering the PDF. The default page
              size is A4.
          paperHeight:
            type: number
            example: 11.69
            description: >-
              Paper height to be used while rendering the PDF. The default page
              size is A4.
          preferCssPageSize:
            type: boolean
            description: >-
              Define whether to prefer page size as defined by CSS (default false)
            default: false
          printBackground:
            type: boolean
            description: >-
              Print the background graphics (default false)
            default: false
          landscape:
            type: boolean
            example: true
            default: false
            description: >-
              The default orientation for rendering the page is "portrait" mode.
              By sending "landscape" parameter, you can ask the output to be
              landscape.
          scale:
            type: number
            minimum: 0.1
            maximum: 2.0
            example: 1.5
            description: >-
              The scale of the page rendering
            default: 1.0
          waitDelay:
            type: string
            example: 5s
            description: >-
              When the page relies on JavaScript for rendering, and you don't have access to the page's code,
              you may want to wait a certain amount of time to make sure Chromium has fully rendered the page
              you're trying to generate.
          waitWindowStatus:
            type: string
            example: done
            description: >-
              If you have access to the page's code, you may set the window status and tell Gotenberg to wait for a specific value.
              For instance
                 await promises()
                 window.status = 'ready'
              Prefer this option over waitDelay.
          extraHttpHeaders:
            type: string
            description: HTTP headers to send by Chromium while loading the HTML document (JSON format)
          nativePageRanges:
            type: string
            example: 1-4
            description: >-
              The page ranges to be converted to PDF for the incoming Office
              documents.
          pdfFormat:
            type: string
            description: >-
              The PDF format of the resulting PDF.
              Caution! You cannot use both nativePdfA1aFormat and pdfFormat form fields.
            example: PDF/A-1a
      - $ref: '#/components/schemas/Encryption'
      required:
        - files
    MarkdownConvertRequestBody:
      title: Markdown Conversion Request Body
      allOf:
      - type: object
        properties:
          files:
            type: array
            items:
              type: string
              format: binary
          marginTop:
            type: number
            example: 0
            default: 1
            description: Top margin for the page in inches.
          marginBottom:
            type: number
            example: 0
            default: 1
            description: Bottom margin for the page in inches.
          marginLeft:
            type: number
            example: 0
            default: 1
            description: Left margin for the page in inches.
          marginRight:
            type: number
            example: 0
            default: 1
            description: Right margin for the page in inches.
          paperWidth:
            type: number
            example: 8.27
            description: >-
              Paper width to be used while rendering the PDF. The default page
              size is A4.
          paperHeight:
            type: number
            example: 11.69
            description: >-
              Paper height to be used while rendering the PDF. The default page
              size is A4.
          preferCssPageSize:
            type: boolean
            description: >-
              Define whether to prefer page size as defined by CSS (default false)
            default: false
          printBackground:
            type: boolean
            description: >-
              Print the background graphics (default false)
            default: false
          landscape:
            type: boolean
            example: true
            default: false
            description: >-
              The default orientation for rendering the page is "portrait" mode.
              By sending "landscape" parameter, you can ask the output to be
              landscape.
          scale:
            type: number
            minimum: 0.1
            maximum: 2.0
            example: 1.5
            description: >-
              The scale of the page rendering
            default: 1.0
          waitDelay:
            type: string
            example: 5s
            description: >-
              When the page relies on JavaScript for rendering, and you don't have access to the page's code,
              you may want to wait a certain amount of time to make sure Chromium has fully rendered the page
              you're trying to generate.
          waitWindowStatus:
            type: string
            example: done
            description: >-
              If you have access to the page's code, you may set the window status and tell Gotenberg to wait for a specific value.
              For instance
                 await promises()
                 window.status = 'ready'
              Prefer this option over waitDelay.
          extraHttpHeaders:
            type: string
            description: HTTP headers to send by Chromium while loading the HTML document (JSON format)
          nativePageRanges:
            type: string
            example: 1-4
            description: >-
              The page ranges to be converted to PDF for the incoming Office
              documents.
          pdfFormat:
            type: string
            description: >-
              The PDF format of the resulting PDF.
              Caution! You cannot use both nativePdfA1aFormat and pdfFormat form fields.
            example: PDF/A-1a
      - $ref: '#/components/schemas/Encryption'
      required:
        - files
    URLConvertRequestBody:
      title: URL Conversion Request Body
      allOf:
      - type: object
        properties:
          url:
            type: string
            example: 'https://google.com'
          files:
            description: Optional files named header.html and footer.html
            type: array
            items:
              type: string
              format: binary
          marginTop:
            type: number
            example: 0
            default: 1
            description: Top margin for the page in inches.
          marginBottom:
            type: number
            example: 0
            default: 1
            description: Bottom margin for the page in inches.
          marginLeft:
            type: number
            example: 0
            default: 1
            description: Left margin for the page in inches.
          marginRight:
            type: number
            example: 0
            default: 1
            description: Right margin for the page in inches.
          paperWidth:
            type: number
            example: 8.27
            description: >-
              Paper width to be used while rendering the PDF. The default page
              size is A4.
          paperHeight:
            type: number
            example: 11.69
            description: >-
              Paper height to be used while rendering the PDF. The default page
              size is A4.
          preferCssPageSize:
            type: boolean
            description: >-
              Define whether to prefer page size as defined by CSS (default false)
            default: false
          printBackground:
            type: boolean
            description: >-
              Print the background graphics (default false)
            default: false
          landscape:
            type: boolean
            example: true
            default: false
            description: >-
              The default orientation for rendering the page is "portrait" mode.
              By sending "landscape" parameter, you can ask the output to be
              landscape.
          scale:
            type: number
            minimum: 0.1
            maximum: 2.0
            example: 1.5
            description: >-
              The scale of the page rendering
            default: 1.0
          waitDelay:
            type: string
            example: 5s
            description: >-
              When the page relies on JavaScript for rendering, and you don't have access to the page's code,
              you may want to wait a certain amount of time to make sure Chromium has fully rendered the page
              you're trying to generate.
          waitWindowStatus:
            type: string
            example: done
            description: >-
              If you have access to the page's code, you may set the window status and tell Gotenberg to wait for a specific value.
              For instance
                 await promises()
                 window.status = 'ready'
              Prefer this option over waitDelay.
          extraHttpHeaders:
            type: string
            description: HTTP headers to send by Chromium while loading the HTML document (JSON format)
          nativePageRanges:
            type: string
            example: 1-4
            description: >-
              The page ranges to be converted to PDF for the incoming Office
              documents.
          pdfFormat:
            type: string
            description: >-
              The PDF format of the resulting PDF.
              Caution! You cannot use both nativePdfA1aFormat and pdfFormat form fields.
            example: PDF/A-1a
      - $ref: '#/components/schemas/Encryption'
      required:
        - url
...
    Encryption:
      type: object
      properties:
        ownerPassword:
          type: string
          description: 'Encryption: The Password used for permissions. Requires userPassword.'
          example: secureOwnerPassword
        userPassword:
          type: string
          description: 'Encryption: The Password used for access to the PDF. Requires ownerPassword.'
          example: secureUserPassword
        keyLength:
          type: integer
          description: 'Encryption: Keylength for encryption, defaults to 256'
          enum: [40, 128, 256]
          nullable: false
          example: 256
  parameters:
    OutputFilename:
      in: header
      name: Gotenberg-Output-Filename
      description: >-
        By default, the API generates a UUID filename.
        However, you may also specify the filename per request,
        thanks to the Gotenberg-Output-Filename header.
        Caution! The API adds the file extension automatically; you don't have to set it.
      schema:
        type: string
      required: false
    Trace:
      in: header
      name: Gotenberg-Trace
      description: >-
        The trace, or request ID, identifies a request in the logs.

        By default, the API generates a UUID trace for each request.
        However, you may also specify the trace per request, thanks to the Gotenberg-Trace header.
      schema:
        type: string
      required: false
  securitySchemes: { }
  responses:
    SuccessfulPDF:
      description: Resulting PDF file from the conversion.
      content:
        application/pdf:
          schema:
            type: string
            format: binary
...